REFA: split query/mod.rs
